Refactor fretboard lookup
